requirements_video_captioner.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.15.0
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0 